Read performance tuning
This change includes the following optimizations:
Router:
1. Using client pool inside Router to reduce lock contention.
Previously, there is only one http aysnc client, which is maintaining a
global connection pool for all the requests, and all the
request-connection operations and release-connection operations will be
blocked by the same mutex. By using a client pool, the lock contention
gets reduced;
2. Doubled the netty IO thread num of Router Server.
The reason is that at one momemont, only one kind of protocol being used
in prod;
3. Reduced the info message logging when there are a big offset lag;
Too much logging will hurt performance greatly;
4. Removed the long-tail retry logic since we don't have a proper way to
enable it since there is only one global config for all kinds of
requests;
5. Updated VeniceRouter to only open listening port when everything is
ready;

Storage Node:
1. Updated bdb cache to use shared cache, and the cache size will up to
60% of JVM size by default;
2. Doubled netty IO thread number of ListenerService;
3. Updated offset lookup to use the in-memory offset data;
4. Updated multi-get request handling logic to execute lookup in
parallel;
5. Updated bdb query latency metric to use 'ms' since the current tehuti
percentile couldn't handle big value properly;
6. Updated 'AggVersionedDIVStats' to warm up per store DIV stats
gradually since the pre-load is taking too long time, which will cause
deployment timeout (in dev cluster, there are more than 2000 stores, and
the deployment also times out every time);

Test setup:
1. Store with 500m entries and each record is about 200 bytes, in total,
the bdb data size is about 125GB in each storage node;
2. 1 Router + 3 Storage Node;
3. 4 Client, and each with 40 single-get threads;

Before:
1. Each Router could handle around 21k qps with latency:
3ms(50%)/7ms(95%)/10ms(99%);

After:
1. Each Router could handle around 37k qps with latency:
2ms(50%)/6ms(95%)/10ms(99%);

For multi-get, it is a little complicated since Router is using greedy
algorithm, which means all the requests will be sent to one single
storage node. And we can compare the performance difference once this
change gets deployed in EI;

For Storage node, right now, each one could handle around 25k~35k qps
with reasonable latency: 2ms(50%)/4ms(95%)6ms(99%);

TODO:
1. Need to execute perf test with more scenarios, such as stores with
various size/various key/value size/various access pattern/ w/(o)
offline data push;
2. Tune GC in Router and Storage Node;

RB=1106788
G=voldemort-reviewers
R=mwise,sdwu,fvillega,chgao,yayan
A=fvillega,yayan

+++ venice-server/src/main/java/com/linkedin/venice/listener/response/ReadResponse.java
@@ -4,13 +4,13 @@ package com.linkedin.venice.listener.response;
  * This class is used to store common fields shared by various read responses.
  */
 public abstract class ReadResponse {
-  private long bdbQueryLatency;
+  private double bdbQueryLatency;
 
-  public void setBdbQueryLatency(long latency) {
+  public void setBdbQueryLatency(double latency) {
     this.bdbQueryLatency = latency;
   }
 
-  public long getBdbQueryLatency() {
+  public double getBdbQueryLatency() {
     return this.bdbQueryLatency;
   }
 }
